# Improve performance of external aggregation with a lot of temporary f…