…iles

porting clickhouse/clickhouse#55489


I try this SQL `SELECT number, count() FROM numbers_mt(5000) GROUP BY number `

```markdown
| Implementation           | Elapsed Time | Processed Rows | Data Processed | Rows/s          | Data/s        |
|--------------------------|--------------|----------------|----------------|-----------------|---------------|
| Proton before this PR    | 0.007 sec    | 5.00 thousand  | 40.00 KB       | 665.96 thousand | 5.33 MB/s     |
| Proton after this PR     | 0.004 sec    | 5.00 thousand  | 40.00 KB       | 1.12 million    | 8.96 MB/s     |
| ClickHouse latest        | 0.003 sec    | 5.00 thousand  | 40.00 KB       | 1.97 million    | 15.78 MB/s    |
```




proton pipeline:
```
:) explain pipeline SELECT number, count() FROM numbers_mt(5000) GROUP BY number 

EXPLAIN PIPELINE
SELECT 
  number, count()
FROM 
  numbers_mt(5000)
GROUP BY 
  number

Query id: f25e0897-b550-476b-bfaa-133582a10464

┌─explain─────────────────┐
│ (Expression)            │
│ ExpressionTransform     │
│   (Aggregating)         │
│   AggregatingTransform  │
│     (Expression)        │
│     ExpressionTransform │
│       (ReadFromStorage) │
│       Limit             │
│         Numbers 0 → 1   │
└─────────────────────────┘

9 rows in set. Elapsed: 0.004 sec. 

:) 
```


clickhouse pipeline:
```
:) explain pipeline  SELECT number, count() FROM numbers_mt(5000) GROUP BY number

EXPLAIN PIPELINE
SELECT
    number,
    count()
FROM numbers_mt(5000)
GROUP BY number

Query id: 5ad3231d-0e56-4e04-adec-49a2beaaa0c8

┌─explain───────────────────┐
│ (Expression)              │
│ ExpressionTransform × 8   │
│   (Aggregating)           │
│   Resize 1 → 8            │
│     AggregatingTransform  │
│       (Expression)        │
│       ExpressionTransform │
│         (ReadFromStorage) │
│         Limit             │
│           Numbers 0 → 1   │
└───────────────────────────┘

10 rows in set. Elapsed: 0.002 sec. 

:) 

```